fix: store term scores as double instead of int32
The Yomitan schema types score as a JSON number and the parser already
reads a double, but the importer narrowed it with static_cast<int32_t>
and the query read it back as int32. Fractional scores collapsed (1.5
and 1.25 tied), and anything outside int32 was undefined behaviour in the
cast (INT32_MIN on x86-64), so those terms sorted below everything.

New imports write the score as an 8-byte double and mark the directory
.hoshidicts_5 (no dict.zstd) or .hoshidicts_6 (dict.zstd), mirroring the
existing _3/_4 pair so a marker still says whether dict.zstd must exist.
_3/_4 dictionaries keep loading and keep reading an int32 score.

TermResult::score, RawTerm::score and hd_term_result.score become double.

+++ include/hoshidicts_c.h
@@ -91,7 +91,7 @@ typedef struct hd_term_result {
   hd_str expression;
   hd_str reading;
   hd_str rules;
-  int32_t score;
+  double score;
   const hd_glossary_entry* glossaries;
   size_t glossaries_count;
   const hd_frequency_entry* frequencies;
